+++ statistics/handle/handletest/lockstats/lock_partition_stats_test.go
@@ -23,7 +23,6 @@ import (
 	"github.com/pingcap/tidb/domain"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/parser/model"
-	"github.com/pingcap/tidb/statistics/handle"
 	"github.com/pingcap/tidb/testkit"
 	"github.com/stretchr/testify/require"
 )
@@ -448,7 +447,7 @@ func TestNewPartitionShouldBeLockedIfWholeTableLocked(t *testing.T) {
 	tk.MustExec("insert into t(a, b) values(21,'a')")
 	tk.MustExec("insert into t(a, b) values(22,'b')")
 	// Dump stats delta to KV.
-	require.Nil(t, h.DumpStatsDeltaToKV(handle.DumpAll))
+	require.Nil(t, h.DumpStatsDeltaToKV(true))
 	// Check the mysql.stats_table_locked is updated correctly.
 	// And the new partition is locked.
 	rows = tk.MustQuery("select count, modify_count, table_id from mysql.stats_table_locked order by table_id").Rows()
@@ -489,7 +488,7 @@ func TestUnlockSomePartitionsWouldUpdateGlobalCountCorrectly(t *testing.T) {
 	require.Equal(t, int64(0), tblStats.RealtimeCount)
 
 	// Dump stats delta to KV.
-	require.Nil(t, h.DumpStatsDeltaToKV(handle.DumpAll))
+	require.Nil(t, h.DumpStatsDeltaToKV(true))
 	// Check the mysql.stats_table_locked is updated correctly.
 	rows := tk.MustQuery("select count, modify_count, table_id from mysql.stats_table_locked order by table_id").Rows()
 	require.Len(t, rows, 2)
