+++ statistics/handle/updatetest/BUILD.bazel
@@ -16,8 +16,8 @@ go_test(
         "//sessionctx",
         "//sessionctx/variable",
         "//statistics",
-        "//statistics/handle",
         "//statistics/handle/autoanalyze",
+        "//statistics/handle/usage",
         "//statistics/handle/util",
         "//testkit",
         "//testkit/testsetup",
