+++ statistics/handle/usage/BUILD.bazel
@@ -10,13 +10,16 @@ go_library(
     importpath = "github.com/pingcap/tidb/statistics/handle/usage",
     visibility = ["//visibility:public"],
     deps = [
+        "//infoschema",
         "//parser/model",
         "//parser/mysql",
         "//sessionctx",
         "//sessionctx/variable",
         "//statistics",
+        "//statistics/handle/storage",
         "//statistics/handle/util",
         "//types",
+        "//util",
         "//util/logutil",
         "//util/sqlexec",
         "@com_github_pingcap_errors//:errors",
