+++ statistics/handle/usage/predicate_column.go
@@ -32,17 +32,34 @@ import (
 
 // statsUsageImpl implements utilstats.StatsUsage.
 type statsUsageImpl struct {
-	pool utilstats.SessionPool
+	pool            utilstats.SessionPool
+	tblInfo         utilstats.TableInfoGetter
+	statsCache      utilstats.StatsCache
+	statsHis        utilstats.StatsHistory
+	getLockedTables func(tableIDs ...int64) (map[int64]struct{}, error) // TODO: use an interface instead of a method pointer
 
 	// idxUsageListHead contains all the index usage collectors required by session.
 	idxUsageListHead *SessionIndexUsageCollector
+
+	// SessionStatsList contains all the stats collector required by session.
+	*SessionStatsList
 }
 
 // NewStatsUsageImpl creates a utilstats.StatsUsage.
-func NewStatsUsageImpl(pool utilstats.SessionPool) utilstats.StatsUsage {
+func NewStatsUsageImpl(pool utilstats.SessionPool,
+	tblInfo utilstats.TableInfoGetter,
+	statsCache utilstats.StatsCache,
+	statsHis utilstats.StatsHistory,
+	getLockedTables func(tableIDs ...int64) (map[int64]struct{}, error),
+) utilstats.StatsUsage {
 	return &statsUsageImpl{
 		pool:             pool,
+		tblInfo:          tblInfo,
+		statsCache:       statsCache,
+		statsHis:         statsHis,
+		getLockedTables:  getLockedTables,
 		idxUsageListHead: newSessionIndexUsageCollector(nil),
+		SessionStatsList: NewSessionStatsList(),
 	}
 }
